Move the serve command to the serve package
